# add a performance testing script