## Summary

<!--
 Ideally, there is an attached Linear ticket that will describe the "why".

 If relevant, use this section to call out any additional information you'd like to _highlight_ to the reviewer.
-->

This PR is a quick script to seed a clickhouse database for the purposes of benchmarking. The reason for this is that our production database is around ~80mb and doesn't give us a lot of confidence of how our queries will perform when we release this out to the wild.

Usage:

```bash
CLICKHOUSE_ADDRESS=GET_FROM_DOPPLER CLICKHOUSE_PASSWORD=GET_FROM_DOPPLER go run backend/clickhouse/performance/main.go
```

This script will insert millions of rows into a the `benchmarks.logs` table (`benchmarks` is the name of the db).

## How did you test this

```sql
select concat(database, '.', table)                         as table,
       formatReadableSize(sum(bytes))                       as size,
       sum(rows)                                            as rows,
       max(modification_time)                               as latest_modification,
       sum(bytes)                                           as bytes_size,
       any(engine)                                          as engine,
       formatReadableSize(sum(primary_key_bytes_in_memory)) as primary_keys_size
from system.parts
where active
and table == 'benchmarks.logs'
group by database, table
order by bytes_size desc;
```

table | size | rows | latest_modification | bytes_size | engine | primary_keys_size
-- | -- | -- | -- | -- | -- | --
benchmarks.logs | 1.52 GiB | 233286669 | 2023-02-07 01:45:34 | 1083089516 | ReplicatedMergeTree | 380.76 KiB

Per the above, **`benchmarks.logs` is 1.52 GB** which is a more realistic data set.

### Get last 100 logs for a given project 🟡 

```sql
SELECT * from benchmarks.logs
WHERE ProjectId = 1
ORDER BY Timestamp DESC
LIMIT 100

// Elapsed: 7.401s Read: 37,634,048 rows (4.92 GB)
```

### Get last 100 logs with a time interval 🟢 

```sql
SELECT * from benchmarks.logs
WHERE ProjectId = 1
AND Timestamp < now()
AND Timestamp >= (now() - toIntervalDay(1))
ORDER BY Timestamp DESC
LIMIT 100

// Elapsed: 1.580s Read: 12,500,992 rows (1.63 GB)
```

### Get all facets for a project 🟢 

```sql
SELECT arrayJoin(LogAttributes.keys) as keys, count() as cnt FROM benchmarks.logs
WHERE ProjectId = 1
GROUP BY keys
ORDER BY cnt DESC;

// Elapsed: 1.825s Read: 37,634,048 rows (602.37 MB)
```

### Get all facets for a project with a time interval 🟢 

```sql
SELECT arrayJoin(LogAttributes.keys) as keys, count() as cnt FROM benchmarks.logs
WHERE ProjectId = 1
AND Timestamp < now()
AND Timestamp >= (now() - toIntervalDay(1))
GROUP BY keys
ORDER BY cnt DESC;

// Elapsed: 0.817s Read: 12,500,992 rows (300.10 MB)
```

### Find all logs given a log attribute 🟡 

```sql
SELECT * from benchmarks.logs
WHERE ProjectId = 1 AND LogAttributes['key7'] = 'val4'
order by Timestamp DESC
LIMIT 1000

// Elapsed: 4.539s Read: 37,634,048 rows (4.68 GB)
```

### Casting 🟡 

```sql
SELECT * from benchmarks.logs
WHERE ProjectId = 1 AND LogAttributes['key7'] = 'val4' AND toInt8(LogAttributes['workspace_id']) > 0
order by Timestamp DESC
LIMIT 1000

// Elapsed: 5.322s Read: 37,634,048 rows (4.67 GB)
```

### Casting with a wildcard 🟡 

```sql
SELECT * from benchmarks.logs
WHERE ProjectId = 1
AND LogAttributes['service_name'] LIKE '%middleware%'
AND toInt8(LogAttributes['workspace_id']) < 99
AND toInt8(LogAttributes['user_id']) > 1
ORDER BY Timestamp DESC
LIMIT 1000

// Elapsed: 7.540s Read: 37,634,048 rows (4.92 GB)
```

### Casting with a wildcard and a time interval 🟡 

```sql
SELECT * from benchmarks.logs
WHERE ProjectId = 1
AND Timestamp < now()
AND Timestamp >= (now() - toIntervalDay(1))
AND LogAttributes['service_name'] LIKE '%middleware%'
AND toInt8(LogAttributes['workspace_id']) < 99
AND toInt8(LogAttributes['user_id']) > 1
ORDER BY Timestamp DESC
LIMIT 1000

// Elapsed: 10.732s Read: 15,562,011 rows (2.13 GB)
```

### Casting with a wildcard, find all facets with a time interval 🔴 

```sql
SELECT arrayJoin(LogAttributes.keys) as keys, count() as cnt FROM benchmarks.logs
WHERE ProjectId = 1
AND Timestamp < now()
AND Timestamp >= (now() - toIntervalDay(1))
AND LogAttributes['service_name'] LIKE '%middleware%'
AND toInt8(LogAttributes['workspace_id']) < 99
AND toInt8(LogAttributes['user_id']) > 1
GROUP BY keys
ORDER BY cnt DESC;

"Code: 33. DB::Exception: Cannot read all data. Bytes read: 0. Bytes expected: 8.: (while reading column LogAttributes.keys): (while reading from part /var/lib/clickhouse/disks/s3disk/store/204/2040e058-ea6e-4fe8-b2b7-2a71cdd01cb0/20230207_0_194_6/ from mark 0 with max_rows_to_read = 16384): While executing MergeTreeThread. (CANNOT_READ_ALL_DATA) (version 22.12.1.21301 (official build))\n"
```

The query times out even if the count is removed.

### Same as previous query but searching on a top level column (`Body`) 🟢 

```sql
SELECT arrayJoin(LogAttributes.keys) as keys, count() as cnt FROM benchmarks.logs
WHERE ProjectId = 1
AND Timestamp < now()
AND Timestamp >= (now() - toIntervalDay(1))
AND Body LIKE '%uvDdArNtsokfhzfqYnzyLZj%'
GROUP BY keys
ORDER BY cnt DESC;

// Elapsed: 0.544s Read: 14,677,275 rows (199.18 MB)
```

The `EXPLAIN` for this query and the previous one are the same. But because we have an index on `Body`, we can filter down the number of rows more easily.

### Adding a materialized column 🟢 

Let's see if we can improve the `Casting with a wildcard and a time interval and `Casting with a wildcard, find all facets with a time inverval` queries by adding a materialized column similar to how [posthog describes](https://posthog.com/blog/clickhouse-materialized-columns). 

```sql
ALTER TABLE logs
ADD COLUMN mat_$service_name
VARCHAR MATERIALIZED LogAttributes['service_name']
```

```sql
/* Backfill data */
OPTIMIZE TABLE logs FINAL 
```

```sql
SELECT * from benchmarks.logs
WHERE ProjectId = 1
AND Timestamp < now()
AND Timestamp >= (now() - toIntervalDay(1))
AND mat_$service_name LIKE '%middleware%'
AND toInt8(LogAttributes['workspace_id']) < 99
AND toInt8(LogAttributes['user_id']) > 1
ORDER BY Timestamp DESC
LIMIT 1000

// Elapsed: 3.621s Read: 15,523,840 rows (2.39 GB)
```
**Previous: 10.7s
Now: ~3.6s**


```sql
SELECT arrayJoin(LogAttributes.keys) as keys, count() as cnt FROM benchmarks.logs
WHERE ProjectId = 1
AND Timestamp < now()
AND Timestamp >= (now() - toIntervalDay(1))
AND mat_$service_name LIKE '%middleware%'
AND toInt8(LogAttributes['workspace_id']) < 99
AND toInt8(LogAttributes['user_id']) > 1
GROUP BY keys
ORDER BY cnt DESC;
// Elapsed: 2.825s Read: 15,562,011 rows (1.65 GB)
```
**Previous: Timedout
Now: ~2.8s**

## TLDR; adding a materialized column speeds up when querying on arbitrary log data (in a clickhouse map)